requirements.txt
pyyaml
requests
bs4
boto3
webexteamssdk
xmltodict
